chore(server): core IoC #95 - getObjectChildrenQueryFactory
