+++ packages/server/modules/core/domain/objects/operations.ts
@@ -66,6 +66,21 @@ export type GetObjectChildren = (params: {
   cursor: string | null
 }>
 
+export type GetObjectChildrenQuery = (params: {
+  streamId: string
+  objectId: string
+  limit?: MaybeNullOrUndefined<number | string>
+  depth?: MaybeNullOrUndefined<number | string>
+  select?: MaybeNullOrUndefined<string[]>
+  cursor?: MaybeNullOrUndefined<string>
+  query?: Array<{ field: string; verb?: string; value: unknown; operator: string }>
+  orderBy?: { field: keyof SpeckleObject; direction: 'asc' | 'desc' }
+}) => Promise<{
+  objects: Omit<SpeckleObject, 'totalChildrenCountByDepth' | 'streamId'>[]
+  cursor: string | null
+  totalCount: number
+}>
+
 export type CreateObject = (params: {
   streamId: string
   object: RawSpeckleObject
