+++ packages/server/modules/core/tests/objects.spec.js
@@ -7,7 +7,7 @@ const { cloneDeep, times, random, padStart } = require('lodash')
 const { beforeEachContext } = require('@/test/hooks')
 const { getAnIdForThisOnePlease } = require('@/test/helpers')
 
-const { getObjects, getObjectChildrenQuery } = require('../services/objects')
+const { getObjects } = require('../services/objects')
 const {
   getStreamFactory,
   createStreamFactory
@@ -83,7 +83,8 @@ const {
   storeObjectsIfNotFoundFactory,
   getFormattedObjectFactory,
   getObjectChildrenStreamFactory,
-  getObjectChildrenFactory
+  getObjectChildrenFactory,
+  getObjectChildrenQueryFactory
 } = require('@/modules/core/repositories/objects')
 
 const sampleCommit = JSON.parse(`{
@@ -188,6 +189,7 @@ const createObjects = createObjectsFactory({
 const getObject = getFormattedObjectFactory({ db })
 const getObjectChildrenStream = getObjectChildrenStreamFactory({ db })
 const getObjectChildren = getObjectChildrenFactory({ db })
+const getObjectChildrenQuery = getObjectChildrenQueryFactory({ db })
 
 describe('Objects @core-objects', () => {
   const userOne = {
